Added Weapon Swing - see description for details
Unity is not able to shift a polygon collision to match an animation. If we were to reuse the same collider, we could create our own animation and rotate it frame by frame ourselves. Otherwise, we would need to have a different polygon collider for each frame that we activate/deactivate as necessary, like this: http://answers.unity3d.com/questions/372252/how-can-i-do-frame-by-frame-hitbox-control-for-a-2.html