Can set outline color. Hitting an enemy with the same outline color does not damage enemy, and knocks you back (no damage)
Moved GetColor to GameController.cs (we really need to update that since we've changed so much, but i'll defer to Persia when she implements gauges). There are some "glowing" enemies to play around with; they are described as glowing because the shaders are taken from https://github.com/Elringus/SpriteGlow, and we can try to emulate if we add in KinoBloom (I think). However, there seems to be problems with Android build so that may be put on the backburner